server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "googlemaps>=4.10.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
